# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Biglands API
  description: |
    Biglands — Bất động sản CHDV platform. API for managing property listings,
    deposits, deal lifecycle, user management, approvals, and notifications.

    **Base URL**: `/api/v1`
    **Auth**: Bearer JWT (24h expiry)
    **Pagination**: offset-based (`page`, `size`)
  version: "1.0.0"
servers:
  - url: /api/v1
    description: Biglands API v1
paths:
...
  /listings:
    post:
      tags: [Listings]
      summary: Create a new listing (Agent/Admin)
      operationId: createListing
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingRequest"
      responses:
        "201":
          description: Listing created (DRAFT or PENDING_APPROVAL depending on save vs submit)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    get:
      tags: [Listings]
      summary: Browse listings with search, filters, and pagination
      operationId: listListings
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - name: q
          in: query
          schema:
            type: string
            maxLength: 200
          description: Search term (matches code, title, description, address, ward, district)
          required: false
        - name: transactionType
          in: query
          schema:
            $ref: "#/components/schemas/TransactionType"
          required: false
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ListingStatus"
          required: false
          description: "Default: CON_HANG, DA_COC (visible listings only)"
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, hot, pinned]
          description: Filter tab
          required: false
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, price, viewCount]
          default: createdAt
          required: false
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
          default: desc
          required: false
      responses:
        "200":
          description: Paginated listing list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingListResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /listings/{id}:
    get:
      tags: [Listings]
      summary: Get listing detail with full data
      operationId: getListing
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      responses:
        "200":
          description: Full listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Listings]
      summary: Edit listing (owner only)
      operationId: updateListing
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingRequest"
      responses:
        "200":
          description: Listing updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Listings]
      summary: Delete a DRAFT listing (owner only)
      operationId: deleteListing
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      responses:
        "204":
          description: Listing permanently deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
          description: "Only DRAFT listings can be deleted (LST-I06)"

  # ──────────────────────────────────────────────
  #  LISTING ACTIONS
  # ──────────────────────────────────────────────
  /listings/{id}/submit:
    post:
      tags: [Listings]
      summary: Submit DRAFT listing for approval
      operationId: submitListing
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      responses:
        "200":
          description: Listing moved to PENDING_APPROVAL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          $ref: "#/components/responses/ValidationError"
          description: "Missing required fields or images (LST-C03)"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "Listing is not in DRAFT status"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /listings/{id}/withdraw:
    post:
      tags: [Listings]
      summary: Withdraw CON_HANG listing back to DRAFT
      operationId: withdrawListing
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      responses:
        "200":
          description: Listing returned to DRAFT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "Listing status is not CON_HANG"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ──────────────────────────────────────────────
  #  LISTING IMAGES
  # ──────────────────────────────────────────────
  /listings/{id}/images:
    post:
      tags: [Images]
      summary: Upload an image to a listing
      operationId: uploadImage
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, WEBP; max 10MB)
              required: [file]
      responses:
        "201":
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingImage"
        "400":
          $ref: "#/components/responses/ValidationError"
          description: "Max 20 images per listing (BR-014) or invalid file type"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Images]
      summary: Reorder images for a listing
      operationId: reorderImages
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderImagesRequest"
      responses:
        "200":
          description: Images reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListingImage"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /listings/{listingId}/images/{imageId}:
    delete:
      tags: [Images]
      summary: Delete an image from a listing
      operationId: deleteImage
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
        - $ref: "#/components/parameters/imageId"
      responses:
        "204":
          description: Image deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /listings/{listingId}/images/{imageId}/primary:
    put:
      tags: [Images]
      summary: Set an image as the primary (cover) image
      operationId: setPrimaryImage
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
        - $ref: "#/components/parameters/imageId"
      responses:
        "200":
          description: Primary image set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingImage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────────
  #  DEAL EVENTS
  # ──────────────────────────────────────────────
  /listings/{id}/deal-events/deposit:
    post:
      tags: [Deal Events]
      summary: Report a customer deposit (Agent)
      operationId: reportDeposit
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportDepositRequest"
      responses:
        "201":
          description: Deposit reported — pending approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealEvent"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "Listing not in CON_HANG status or duplicate active deposit (DE-I02)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /listings/{id}/deal-events/closure:
    post:
      tags: [Deal Events]
      summary: Report deal closure (Agent)
      operationId: reportClosure
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportClosureRequest"
      responses:
        "201":
          description: Closure reported — pending approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealEvent"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "Listing not in DA_COC status"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /listings/{id}/deal-events/cancellation:
    post:
      tags: [Deal Events]
      summary: Report deposit cancellation (Agent)
      operationId: reportCancellation
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCancellationRequest"
      responses:
        "201":
          description: Cancellation reported — pending approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealEvent"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "Listing not in DA_COC status"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /listings/{id}/deal-events/sold-out:
    post:
      tags: [Deal Events]
      summary: Report listing as sold out (Agent)
      operationId: reportSoldOut
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/listingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportSoldOutRequest"
      responses:
        "201":
          description: Sold-out reported — pending approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealEvent"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "Listing not in CON_HANG status"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ──────────────────────────────────────────────
  #  APPROVALS
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer token. Issued via POST /auth/login. Expires in 24h."
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
      required: false
    size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
      required: false
...
    listingId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Listing UUID
    imageId:
      name: imageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Image UUID
...
  responses:
    ValidationError:
      description: "Validation error — field-level or business-rule violation"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: "Missing, invalid, or expired authentication token"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: "Authenticated but insufficient permissions or not the resource owner"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: "Requested resource was not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Conflict:
      description: "Resource conflict or invalid state transition"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    TransactionType:
      type: string
      enum: [SANG_NHUONG, CHO_THUE, BAN]
      description: "Transaction type for a listing."
    ListingStatus:
      type: string
      enum: [DRAFT, PENDING_APPROVAL, CON_HANG, DA_COC, HET_HANG, DA_CHOT, HUY_COC, QUA_HAN]
      description: "Listing status machine. See domain-model.md §11 for state transitions."
...
    Listing:
      type: object
      required:
        - id
        - code
        - transactionType
        - propertyType
        - description
        - price
        - commissionType
        - commissionValue
        - areaWidth
        - areaLength
        - totalArea
        - numRooms
        - numBathrooms
        - numFloors
        - streetName
        - houseNumber
        - address
        - ward
        - district
        - city
        - ownerPhone
        - status
        - createdById
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 20
          description: "Auto-generated: YYMMDD + random digits (LST-C01)"
        transactionType:
          $ref: "#/components/schemas/TransactionType"
        propertyType:
          $ref: "#/components/schemas/PropertyType"
        title:
          type: string
          maxLength: 500
          nullable: true
          description: "Optional title with optional prefix (NNC, MT, HXH, CHDV)"
        description:
          type: string
          description: Property description
        price:
          type: number
          description: Price in VND
        commissionType:
          $ref: "#/components/schemas/CommissionType"
        commissionValue:
          type: number
          description: "Percentage (e.g., 1.5) or fixed VND amount"
        areaWidth:
          type: number
          format: float
          description: Frontage width in meters
        areaLength:
          type: number
          format: float
          description: Property length in meters
        totalArea:
          type: number
          format: float
          description: "Total floor area in m² (may be computed)"
        numRooms:
          type: integer
          minimum: 0
          description: Number of bedrooms
        numBathrooms:
          type: integer
          minimum: 0
          description: Number of bathrooms/toilets
        numFloors:
          type: integer
          minimum: 0
          description: Number of floors
        streetName:
          type: string
          maxLength: 255
        houseNumber:
          type: string
          maxLength: 50
        address:
          type: string
          maxLength: 500
          description: "Full composite address"
        ward:
          type: string
          maxLength: 100
        district:
          type: string
          maxLength: 100
        city:
          type: string
          maxLength: 100
          default: "Hồ Chí Minh"
        latitude:
          type: number
          format: float
          nullable: true
        longitude:
          type: number
          format: float
          nullable: true
        label:
          type: string
          maxLength: 100
          nullable: true
          description: "e.g., Thang máy, Nhà mới, Vị trí đẹp"
        furnishing:
          type: string
          maxLength: 500
          nullable: true
          description: Furniture description
        frontageType:
          type: string
          maxLength: 100
          nullable: true
          description: "Mặt tiền / Hẻm"
        legalStatus:
          type: string
          maxLength: 500
          nullable: true
          description: Legal documentation
        direction:
          type: string
          maxLength: 50
          nullable: true
          description: "Hướng — e.g., Đông, Tây, Nam, Bắc"
        roadWidth:
          type: string
          maxLength: 50
          nullable: true
          description: Alley/road width
        ownerPhone:
          type: string
          maxLength: 20
          description: Property owner's contact phone
        videoUrl:
          type: string
          format: uri
          maxLength: 500
          nullable: true
          description: "YouTube video link (max 1 per listing — BR-014)"
        status:
          $ref: "#/components/schemas/ListingStatus"
        isHot:
          type: boolean
          default: false
          description: Promoted to Hot by Admin
        hotOrder:
          type: integer
          nullable: true
          description: Display order in Hot section
        viewCount:
          type: integer
          default: 0
        createdById:
          type: string
          format: uuid
          description: Agent who created the listing
        approvedById:
          type: string
          format: uuid
          nullable: true
          description: Approver who approved
        approvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateListingRequest:
      type: object
      required:
        - transactionType
        - propertyType
        - description
        - price
        - commissionType
        - commissionValue
        - areaWidth
        - areaLength
        - numRooms
        - numBathrooms
        - numFloors
        - streetName
        - houseNumber
        - ward
        - district
        - ownerPhone
      properties:
        transactionType:
          $ref: "#/components/schemas/TransactionType"
        propertyType:
          $ref: "#/components/schemas/PropertyType"
        title:
          type: string
          maxLength: 500
        description:
          type: string
          minLength: 1
        price:
          type: number
          minimum: 0
          description: Price in VND
        commissionType:
          $ref: "#/components/schemas/CommissionType"
        commissionValue:
          type: number
          minimum: 0
          description: "Percentage or fixed VND (BR-008)"
        areaWidth:
          type: number
          format: float
          minimum: 0
          description: Frontage width in meters
        areaLength:
          type: number
          format: float
          minimum: 0
          description: Property length in meters
        numRooms:
          type: integer
          minimum: 0
          default: 0
        numBathrooms:
          type: integer
          minimum: 0
          default: 0
        numFloors:
          type: integer
          minimum: 0
          default: 0
        streetName:
          type: string
          maxLength: 255
          minLength: 1
        houseNumber:
          type: string
          maxLength: 50
          minLength: 1
        ward:
          type: string
          maxLength: 100
          minLength: 1
          description: Must be valid for the selected district (BR-013)
        district:
          type: string
          maxLength: 100
          minLength: 1
          description: Must be valid for the selected city (BR-013)
        city:
          type: string
          maxLength: 100
          default: "Hồ Chí Minh"
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        label:
          type: string
          maxLength: 100
        furnishing:
          type: string
          maxLength: 500
        frontageType:
          type: string
          maxLength: 100
        legalStatus:
          type: string
          maxLength: 500
        direction:
          type: string
          maxLength: 50
        roadWidth:
          type: string
          maxLength: 50
        ownerPhone:
          type: string
          maxLength: 20
          minLength: 1
          description: Property owner's phone
        videoUrl:
          type: string
          format: uri
          maxLength: 500
        action:
          type: string
          enum: [save, submit]
          default: save
          description: "save → DRAFT; submit → PENDING_APPROVAL"
    UpdateListingRequest:
      type: object
      description: "Partial update — only provided fields are changed. Changing price, areaWidth, areaLength, or totalArea on CON_HANG triggers re-approval (LST-I05)."
      properties:
        propertyType:
          $ref: "#/components/schemas/PropertyType"
        title:
          type: string
          maxLength: 500
          nullable: true
        description:
          type: string
        price:
          type: number
          minimum: 0
        commissionType:
          $ref: "#/components/schemas/CommissionType"
        commissionValue:
          type: number
          minimum: 0
        areaWidth:
          type: number
          format: float
          minimum: 0
        areaLength:
          type: number
          format: float
          minimum: 0
        numRooms:
          type: integer
          minimum: 0
        numBathrooms:
          type: integer
          minimum: 0
        numFloors:
          type: integer
          minimum: 0
        streetName:
          type: string
          maxLength: 255
        houseNumber:
          type: string
          maxLength: 50
        ward:
          type: string
          maxLength: 100
        district:
          type: string
          maxLength: 100
        city:
          type: string
          maxLength: 100
        label:
          type: string
          maxLength: 100
          nullable: true
        furnishing:
          type: string
          maxLength: 500
          nullable: true
        frontageType:
          type: string
          maxLength: 100
          nullable: true
        legalStatus:
          type: string
          maxLength: 500
          nullable: true
        direction:
          type: string
          maxLength: 50
          nullable: true
        roadWidth:
          type: string
          maxLength: 50
          nullable: true
        ownerPhone:
          type: string
          maxLength: 20
        videoUrl:
          type: string
          format: uri
          maxLength: 500
          nullable: true
    ListingDetailResponse:
      type: object
      required:
        - listing
        - images
        - dealEvents
        - isPinned
      properties:
        listing:
          $ref: "#/components/schemas/Listing"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ListingImage"
        dealEvents:
          type: array
          items:
            $ref: "#/components/schemas/DealEvent"
        isPinned:
          type: boolean
          description: "Whether current user has pinned this listing"
    ListingListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        pagination:
          $ref: "#/components/schemas/Pagination"
        totalCount:
          type: integer
          description: "Total count of CON_HANG listings across all transaction types (BR-015)"

    # ── ListingImage ───────────────────────────
    ListingImage:
      type: object
      required: [id, listingId, url, order]
      properties:
        id:
          type: string
          format: uuid
        listingId:
          type: string
          format: uuid
        url:
          type: string
          format: uri
          maxLength: 1000
        order:
          type: integer
          minimum: 1
          description: Display order (unique per listing — IMG-C02)
        isPrimary:
          type: boolean
          default: false
          description: "Cover image (at most one per listing — IMG-I01)"
    ReorderImagesRequest:
      type: object
      required: [order]
      properties:
        order:
          type: array
          items:
            type: string
            format: uuid
          description: "Array of image UUIDs in desired display order (positional)"

    # ── DealEvent ──────────────────────────────
...
    ReportClosureRequest:
      type: object
      properties:
        notes:
          type: string
    ReportCancellationRequest:
      type: object
      required: [notes]
      properties:
        notes:
          type: string
          minLength: 1
          description: "Cancellation reason is required (DE-C06)"
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: "Error code — e.g., VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, ALREADY_PROCESSED, INVALID_STATUS_TRANSITION"
        message:
          type: string
          description: Human-readable error description
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: "Field path for validation errors"
              message:
                type: string
                description: "Field-level error message"
              code:
                type: string
                description: "Field-level error code: required, minLength, maxLength, pattern, minimum, maximum, invalid, duplicate, conflict"